# specs/001-crm-intelligent-analytics/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CRM智能分析系统一期接口
  version: 1.2.0
  description: >
    面向一期企业微信 AI 问数、Web 智能分析工作台与智能合同审核模块的接口契约，
    覆盖自然语言问数、企业微信消息接入、查询模板、最近查询、结果查询、
    合同审核任务、审核产物下载、导出控制、治理配置、审计检索与销售日报编排。
servers:
  - url: https://crm-analysis.internal.example.com
    description: 内网生产环境示例地址
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/wecom/messages:
    post:
      tags: [WecomBot]
      summary: 接收企业微信机器人消息
      operationId: receiveWecomMessage
      security:
        - wecomSignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WecomIncomingMessageRequest'
      responses:
        '202':
          description: 消息已接收，系统开始处理或继续当前会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WecomIncomingMessageResponse'
        '403':
          description: 当前用户无权使用企业微信问数能力
        '409':
          description: 会话状态冲突或当前请求需要排队
  /api/v1/wecom/sessions/{sessionId}:
    get:
      tags: [WecomBot, Session]
      summary: 查询企业微信会话状态
      operationId: getWecomSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 当前会话详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySessionResponse'
        '404':
          description: 会话不存在
  /api/v1/wecom/sessions/{sessionId}/heartbeat:
    post:
      tags: [Session]
      summary: 上报企业微信会话心跳
      operationId: reportWecomSessionHeartbeat
      security:
        - wecomSignatureAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionHeartbeatRequest'
      responses:
        '200':
          description: 心跳已记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionHeartbeatResponse'
        '404':
          description: 会话不存在
...
  /api/v1/analysis/histories:
    get:
      tags: [QueryAsset]
      summary: 获取当前用户最近查询列表
      operationId: listRecentQueries
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: 最近查询分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentQueryListResponse'
  /api/v1/analysis/histories/{historyId}/rerun:
    post:
      tags: [QueryAsset]
      summary: 按当前权限重新运行最近查询
      operationId: rerunRecentQuery
      parameters:
        - $ref: '#/components/parameters/HistoryId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RerunRecentQueryRequest'
      responses:
        '201':
          description: 最近查询已重新执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisQueryResponse'
        '403':
          description: 当前用户无权重新执行该记录
        '404':
          description: 最近查询记录不存在
...
  /api/v1/governance/query-templates:
    get:
      tags: [Governance]
      summary: 获取治理端查询模板列表
      operationId: listGovernanceQueryTemplates
      responses:
        '200':
          description: 模板列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryTemplateListResponse'
    post:
      tags: [Governance]
      summary: 新增查询模板
      operationId: createGovernanceQueryTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQueryTemplateRequest'
      responses:
        '201':
          description: 模板创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryTemplateResponse'
        '400':
          description: 模板内容不合法
        '403':
          description: 当前用户不是管理员
  /api/v1/governance/query-templates/{templateId}:
    put:
      tags: [Governance]
      summary: 更新指定查询模板
      operationId: updateGovernanceQueryTemplate
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQueryTemplateRequest'
      responses:
        '200':
          description: 模板更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryTemplateResponse'
        '400':
          description: 模板内容不合法
        '403':
          description: 当前用户不是管理员
        '404':
          description: 模板不存在
    delete:
      tags: [Governance]
      summary: 删除指定查询模板
      operationId: deleteGovernanceQueryTemplate
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: 模板删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteQueryTemplateResponse'
        '403':
          description: 当前用户不是管理员
        '404':
          description: 模板不存在
  /api/v1/governance/query-templates/{templateId}/validate:
    post:
      tags: [Governance]
      summary: 校验指定查询模板 SQL
      operationId: validateGovernanceQueryTemplate
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '201':
          description: 模板 SQL 校验结果
  /api/v1/governance/query-templates/{templateId}/preview:
    post:
      tags: [Governance]
      summary: 预览指定查询模板结果
      operationId: previewGovernanceQueryTemplate
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '201':
          description: 模板预览结果
...
  /api/v1/audit-events:
    get:
      tags: [Audit]
      summary: 查询审计事件列表
      operationId: listAuditEvents
      parameters:
        - in: query
          name: actorId
          schema:
            type: string
          description: 按行为人筛选
        - in: query
          name: eventType
          schema:
            $ref: '#/components/schemas/AuditEventType'
          description: 按事件类型筛选
        - in: query
          name: startAt
          schema:
            type: string
            format: date-time
          description: 开始时间
        - in: query
          name: endAt
          schema:
            type: string
            format: date-time
          description: 结束时间
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 审计事件分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventListResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: crm_auth_session
    wecomSignatureAuth:
      type: apiKey
      in: header
      name: X-Wecom-Signature
  parameters:
...
    SessionId:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
      description: 查询会话 ID
    HistoryId:
      in: path
      name: historyId
      required: true
      schema:
        type: string
      description: 最近查询记录 ID
    TemplateId:
      in: path
      name: templateId
      required: true
      schema:
        type: string
      description: 查询模板 ID
...
  schemas:
...
    WecomIncomingMessageRequest:
      type: object
      description: >
        兼容当前本地联调简化报文与企业微信官方风格文本消息。
        一期只支持文本问数，不支持图片、文件或语音消息直接进入正式分析链路。
      properties:
        externalConversationId:
          type: string
          description: 简化联调报文中的外部会话标识
        senderId:
          type: string
          description: 简化联调报文中的发送者标识
        messageId:
          type: string
          description: 简化联调报文中的消息唯一标识
        messageText:
          type: string
          minLength: 1
          maxLength: 1000
          description: 简化联调报文中的文本内容
        msgid:
          type: string
          description: 企业微信官方文本消息的唯一标识
        chattype:
          type: string
          enum: [single, group]
        chatid:
          type: string
        msgtype:
          type: string
          enum: [text, image, mixed, voice, file]
        from:
          type: object
          properties:
            userid:
              type: string
        text:
          type: object
          properties:
 

              type: string
        followUpToSessionId:
          type: string
        sentAt:
          type: string
          format: date-time
    WecomIncomingMessageResponse:
      type: object
      required: [receiptId, status, acceptedAt]
      properties:
        receiptId:
          type: string
        sessionId:
          type: string
        queryId:
          type: string
        status:
          type: string
          description: 企业微信消息当前处理结果；跟进写回相关状态仅在进入受控跟进流程后返回，主题型跟进入口在模板收集中通常返回 `DAILY_REPORT_PROMPTED`，CRM 创建相关状态仅在受控新增客户/商机流程中返回。
          enum: [RETURNED, CLARIFICATION_REQUIRED, QUEUED, BLOCKED, EXPLAINED, DEGRADED, OPPORTUNITY_LOOKUP_RETURNED, ENTITY_LOOKUP_LIST_RETURNED, ENTITY_LOOKUP_DETAIL_RETURNED, FOLLOW_UP_WRITEBACK_PENDING_CONFIRMATION, FOLLOW_UP_WRITEBACK_AWAITING_CONTENT_CONFIRMATION, FOLLOW_UP_WRITEBACK_SUCCEEDED, FOLLOW_UP_WRITEBACK_CANCELLED, FOLLOW_UP_WRITEBACK_FAILED, FOLLOW_UP_SHARE_PENDING_CONFIRMATION, FOLLOW_UP_SHARE_SUCCEEDED, FOLLOW_UP_SHARE_CANCELLED, WECOM_CRM_CREATE_COLLECTING, WECOM_CRM_CREATE_AWAITING_CONFIRMATION, WECOM_CRM_CREATE_SUCCEEDED, WECOM_CRM_CREATE_CANCELLED, WECOM_CRM_CREATE_FAILED]
        acceptedAt:
          type: string
          format: date-time
        deduplicated:
          type: boolean
        queueNotice:
          type: string
        clarificationPrompt:
          type: string
        deliveryStatus:
          type: string
          enum: [PENDING, SENT, RETRYING, FAILED, SKIPPED]
        deliveredBlockCount:
          type: integer
          minimum: 0
        followUpWriteback:
          description: 仅在进入受控跟进写回确认阶段时返回；主题型模板收集阶段、普通问数或拒绝场景不返回该字段。进入该阶段前，服务端已经同时完成 `wecom.followup.writeback` 动作权限校验，以及负责人 / 协作人 / 递归上级链对象级授权校验。
          $ref: '#/components/schemas/WecomFollowUpWritebackPayload'
        crmCreate:
          description: 仅在企业微信受控新增客户或受控新增商机流程中返回；普通问数、拒绝或跟进写回场景不返回该字段。
          $ref: '#/components/schemas/WecomCrmCreatePayload'
    WecomFollowUpWritebackPayload:
      type: object
      required: [id, objectType, objectId, objectTitle, draftContent, status]
      properties:
        id:
          type: string
        objectType:
          type: string
          enum: [Customer, Opportunity]
        objectId:
          type: string
        objectTitle:
          type: string
        customerName:
          type: string
        draftContent:
          type: string
        status:
          $ref: '#/components/schemas/WecomFollowUpWritebackStatus'
        failureReason:
          type: string
        externalRevisitLogId:
          type: string
        writtenAt:
          type: string
          format: date-time
...
    QuerySessionResponse:
      type: object
      required: [sessionId, channel, status, requesterId, lastMessageAt]
      properties:
        sessionId:
          type: string
        channel:
          $ref: '#/components/schemas/ChannelType'
        status:
          $ref: '#/components/schemas/QuerySessionStatus'
        requesterId:
          type: string
        activeQueryId:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        lastHeartbeatAt:
          type: string
          format: date-time
        pendingSequence:
          type: integer
          minimum: 0
        disconnectReason:
          type: string
    SessionHeartbeatRequest:
      type: object
      required: [reportedAt]
      properties:
        reportedAt:
          type: string
          format: date-time
        connectionState:
          type: string
          enum: [ACTIVE, IDLE]
    SessionHeartbeatResponse:
      type: object
      required: [sessionId, status, receivedAt]
      properties:
        sessionId:
          type: string
        status:
          $ref: '#/components/schemas/QuerySessionStatus'
        receivedAt:
          type: string
          format: date-time
...
    AnalysisQueryResponse:
      type: object
      required: [queryId, status, createdAt]
      properties:
        queryId:
          type: string
        status:
          $ref: '#/components/schemas/AnalysisRequestStatus'
        clarificationPrompt:
          type: string
        missingConditions:
          type: array
          items:
            type: string
        sessionId:
          type: string
        queueNotice:
          type: string
        createdAt:
          type: string
          format: date-time
...
    QueryTemplateListResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QueryTemplateResponse'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
...
    DeleteQueryTemplateResponse:
      type: object
      required: [success, templateId]
      properties:
        success:
          type: boolean
        templateId:
          type: string
...
    QueryTemplateResponse:
      type: object
      required: [templateId, name, description, defaultFilters, visibleRoleIds, displayOrder, clickCount7d, hitRatePercent, optimizationStatus, status, updatedAt]
      properties:
        templateId:
          type: string
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        defaultQuestionText:
          type: string
        defaultFilters:
          type: object
          additionalProperties: true
        defaultViewType:
          $ref: '#/components/schemas/ResultViewType'
        queryMode:
          type: string
          enum: [FIXED_SQL]
        sqlText:
          type: string
        sqlVersion:
          type: string
        sourceType:
          type: string
          enum: [GOVERNANCE_CREATED, FREE_QUERY_SAVED, COPIED_FROM_TEMPLATE, LEGACY_MIGRATED]
        sourceQueryId:
          type: string
        sourceTemplateId:
          type: string
        sourceSnapshot:
          type: object
          additionalProperties: true
        scopeMode:
          type: string
          enum: [AUTO_SCOPE, DECLARED_SCOPE]
        parameterSchema:
          type: array
          items:
            type: object
            additionalProperties: true
        renderConfig:
          type: object
          additionalProperties: true
        recommendationReason:
          type: string
        validationSnapshot:
          $ref: '#/components/schemas/QueryTemplateValidationSnapshot'
        lastValidatedAt:
          type: string
          format: date-time
        visibleRoleIds:
          type: array
          items:
            type: string
        ownerUserId:
          type: string
        visibilityType:
          type: string
          enum: [PRIVATE, SHARED]
        displayOrder:
          type: integer
          minimum: 0
        clickCount7d:
          type: integer
          minimum: 0
        usageCountTotal:
          type: integer
          minimum: 0
        lastUsedAt:
          type: string
          format: date-time
        hitRatePercent:
          type: number
          minimum: 0
          maximum: 100
        optimizationStatus:
          $ref: '#/components/schemas/OptimizationStatus'
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        updatedAt:
          type: string
          format: date-time
    RecentQueryListResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecentQueryResponse'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    RerunRecentQueryRequest:
      type: object
      properties:
        overrideQuestionText:
          type: string
          maxLength: 500
        overrideFilters:
          type: object
          additionalProperties: true
        channel:
          $ref: '#/components/schemas/ChannelType'
...
    CreateQueryTemplateRequest:
      type: object
      required: [name, description, defaultQuestionText, defaultFilters, visibleRoleIds, displayOrder, status]
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        defaultQuestionText:
          type: string
        defaultFilters:
          type: object
          additionalProperties: true
        defaultViewType:
          $ref: '#/components/schemas/ResultViewType'
        sqlText:
          type: string
        sqlVersion:
          type: string
        parameterSchema:
          type: array
          items:
            type: object
            additionalProperties: true
        renderConfig:
          type: object
          additionalProperties: true
        visibleRoleIds:
          type: array
          items:
            type: string
        ownerUserId:
          type: string
        visibilityType:
          type: string
          enum: [PRIVATE, SHARED]
        displayOrder:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    UpdateQueryTemplateRequest:
      type: object
      required: [name, description, defaultQuestionText, defaultFilters, visibleRoleIds, displayOrder, status]
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        defaultQuestionText:
          type: string
        defaultFilters:
          type: object
          additionalProperties: true
        defaultViewType:
          $ref: '#/components/schemas/ResultViewType'
        sqlText:
          type: string
        sqlVersion:
          type: string
        parameterSchema:
          type: array
          items:
            type: object
            additionalProperties: true
        renderConfig:
          type: object
          additionalProperties: true
        visibleRoleIds:
          type: array
          items:
            type: string
        ownerUserId:
          type: string
        visibilityType:
          type: string
          enum: [PRIVATE, SHARED]
        usageCountTotal:
          type: integer
          minimum: 0
        lastUsedAt:
          type: string
          format: date-time
        displayOrder:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
...
    AuditEventListResponse:
      type: object
      required: [summary, items, page, pageSize, total]
      properties:
        summary:
          $ref: '#/components/schemas/AuditEventSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...